fastmcp_server/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.5
httpx==0.27.0
pydantic==2.8.2
qdrant-client==1.15.1
google-generativeai==0.7.2
python-dotenv==1.0.1
sqlite-utils==3.36
